\name{mouse}
\docType{data}
\alias{mouse}

\title{Mouse  Dataset}

\description{Mouse (Microtus Pennsylvanicus) Dataset}

\usage{mouse}
\format{A matrix with 104 rows (observed animals) and 5 columns (capture occasions)}

\details{The mouse (Microtus pennsylvanicus)
data were first discussed in Nichols, Pollock and Hines (1984). The original
live-trapping experiment was conducted monthly from June to December, 1980. During
each month, the capture-recapture procedure was repeated for 5 consecutive days.
The detailed data are given in Williams, Nichols and Conroy (2002, pp. 525-528). We
use the data collected in June. A total of 104 distinct mice were caught in the
experiment.
}


\references{
J. D. Nichols, K. H. Pollock and J. E. Hines,
The Use of a Robust Capture-Recapture Design
in Small Mammal Population Studies:
A Field Example with Microtus Pennsylvanicus,
Acta Theriologica, vol. 29. 30:357-365, 1984
}


\examples{
data(mouse)
}

\keyword{Datasets}



