% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-data-documentation.R
\docType{data}
\name{hornedlizard}
\alias{hornedlizard}
\title{Flat-tailed Horned Lizard Dataset}
\format{A matrix with 68 rows (observed lizards) and 14 columns (capture occasions)}
\source{
Royle, J. A. and Young, K. V. (2008)
}
\usage{
data(hornedlizard)
}
\description{
Data from multiple searches for flat-tailed horned lizards (\emph{Phrynosoma mcalli}) on a plot in Arizona, USA.
}
\details{
The flat-tailed horned lizard (\emph{Phrynosoma mcalli}) is a desert lizard found in parts of southwestern Arizona, southeastern California and northern Mexico. There is considerable concern about its conservation status. The species is cryptically colored and has the habit of burying under the sand when approached, making it difficult or impossible to obtain a complete count (Grant and Doherty 2007).

A total of 68 individuals were captured 134 times. Exactly half of the individuals were recaptured exactly only once. This dataset is also includedin the \code{secr} package.
}
\examples{
data(hornedlizard)
str(hornedlizard)

}
\references{
Grant, T. J. and Doherty, P. F. (2007) Monitoring of the flat-tailed horned lizard with methods incorporating detection probability. \emph{Journal of Wildlife Management} \bold{71}, 1050--1056

Royle, J. A. and Young, K. V. (2008) A hierarchical model for spatial capture–recapture data. Ecology 89, 2281–2289.
}
\keyword{Datasets}
